+++ src/Common/CsvTsv/inc/CsvTsv/Table.h
@@ -757,7 +757,7 @@ namespace CsvTsv
         if (static_cast<size_t>(m_cursor + 1) < m_data.size())
         {
             ++m_cursor;
-            this->SetValue(m_data[m_cursor]);
+            this->SetValue(m_data[static_cast<size_t>(m_cursor)]);
             success = true;
         }
